generalize clickable list row, add to folder and teams
